backend_database/requirements.txt
fastapi
uvicorn
pillow
deepface
python-multipart
supabase
python-dotenv
opencv-python-headless
unidecode